move topk capturers to srt/state_capturer/
