+++ test/registered/rl/test_return_routed_experts.py
@@ -9,7 +9,7 @@ import torch
 from torch.nn.utils.rnn import pad_sequence
 
 from sglang.benchmark.utils import download_and_cache_hf_file
-from sglang.srt.layers.moe.routed_experts_capturer import (
+from sglang.srt.state_capturer.routed_experts import (
     extract_routed_experts_from_meta_info,
 )
 from sglang.srt.utils import kill_process_tree
